+++ docs/changelog.md
@@ -6,6 +6,25 @@ All notable changes to Mixtape Society are documented here. We follow [Semantic 
 
 Each release rewinds and fast-forwards your mixtape experience—just like a real cassette! 🎧
 
+## 📦 v0.5.5
+
+<span class="md-tag">Release</span>
+
+**🗓️** 2026-01-09
+
+### ✨ Added
+
+- Sharing through QR codes on browse, editor and playback pages
+- Filtering and sorting on the mixtape browsing page
+
+### 🔧 Changed
+
+- Changed search bar for the mixtape editor
+
+[Changes since last version](https://github.com/mark-me/mixtape-society/compare/v0.5.4...v0.5.5)
+
+---
+
 ## 📦 v0.5.4
 
 <span class="md-tag">Release</span>